Cleaned up and organised files
